Rename config.sample.toml config.toml when running
In the repo it makes sense to present the config file as a sample or
default config, however when templating a config file running in
production, the sample designator doesn't make sense.

This patch will rename it to `config.toml` within the Dockerfile and
instead load from `config.toml` within the main function.

Alternatives considered:
* Renaming to `config.default.toml`
* Providing a command line argument containing a path (This would be in
  addition to the rename"